Documento Resumen desarrollo squeesar

+++ MetodosSQL/Poligono/02 FUNCTION  intersecta_poligono_squeesar.sql
@@ -30,53 +30,54 @@ BEGIN
 	WHERE
 		vigencia = 'S';
 		
-		
-	DELETE FROM 
-		poligonos.poligono_squeesar
-	WHERE
-		id_poligono = nIdPoligono$;
-		
-	sQuery$ := 'SELECT DISTINCT
-					id_squeesar_consolidado as squeesar,
-					id as poligono	
-				FROM
-					squeesar.' || sNombreTabla$ || '
-				INNER JOIN poligonos.poligono ON ST_Intersects (
-					squeesar.' || sNombreTabla$ || '.geom,
-					poligonos.poligono.geom
-					)
-				WHERE
-					ST_isvalid (squeesar.' || sNombreTabla$ || '.geom) = ''t''
-					AND ST_isvalid (poligonos.poligono.geom) = ''t''
-					AND poligonos.id = ' || nIdPoligono$;
-					
-					--Raise notice 'sQuery$ %',sQuery$;
-					
-	FOR cur_Interseccion IN
-		EXECUTE sQuery$
-	LOOP
-
-		INSERT INTO
+	IF NEW.geom <> OLD.geom THEN 	
+		DELETE FROM 
 			poligonos.poligono_squeesar
-			(
-				id_squeesar_consolidado,
-				id_poligono
-			)
-			VALUES
-			(
-				cur_Interseccion.squeesar,
-				cur_Interseccion.poligono
-			);
+		WHERE
+			id_poligono = nIdPoligono$;
 			
-	END LOOP;
-		
-	SELECT 
-		*
-	INTO
-		sResult$
-	FROM 
-		squeesar.agrupar_squeesar(nIdPoligono$);
+		sQuery$ := 'SELECT DISTINCT
+						id_squeesar_consolidado as squeesar,
+						id as poligono	
+					FROM
+						squeesar.' || sNombreTabla$ || '
+					INNER JOIN poligonos.poligono ON ST_Intersects (
+						squeesar.' || sNombreTabla$ || '.geom,
+						poligonos.poligono.geom
+						)
+					WHERE
+						ST_isvalid (squeesar.' || sNombreTabla$ || '.geom) = ''t''
+						AND ST_isvalid (poligonos.poligono.geom) = ''t''
+						AND poligonos.id = ' || nIdPoligono$;
+						
+						--Raise notice 'sQuery$ %',sQuery$;
+						
+		FOR cur_Interseccion IN
+			EXECUTE sQuery$
+		LOOP
+
+			INSERT INTO
+				poligonos.poligono_squeesar
+				(
+					id_squeesar_consolidado,
+					id_poligono
+				)
+				VALUES
+				(
+					cur_Interseccion.squeesar,
+					cur_Interseccion.poligono
+				);
+				
+		END LOOP;
 			
+		SELECT 
+			*
+		INTO
+			sResult$
+		FROM 
+			squeesar.agrupar_squeesar(nIdPoligono$);
+	END IF;
+	
 	RETURN NEW;
 END;
 $BODY$
